Add /published?id= entry point
