+++ src/api/index.js
@@ -6,5 +6,6 @@ apis.push(require('./healthcheck'));
 apis.push(require('./search'));
 apis.push(require('./stats'));
 apis.push(require('./taxonomy'));
+apis.push(require('./published'));
 
 module.exports = apis;
